vcert/v4/pkg: Expose API for refreshing the access token
RefreshAccessToken() api is not exposed in Connector interface
at the moment. Adding the same by introducing a wrapper over
RefreshAccessToken() which returns an interface instead of tpp
specific struct.

Other possible approach can be - moving out the struct
OauthRefreshAccessTokenResponse to endpoint.go file and add
RefreshAccessToken() to Connector interface. But the stuct
is TPP specific so placing it in endpoint pkg does not look
right therefor going with current approac