revert to the name "resolveContacts" and re-introduce getIdentities
The problem with resolvePrefixedUniversals is that it was using
getIdentity, which only works for usernames due to getIdentityMatching.
Since I don't know why the name check is performed in
getIdentityMatching, I prefer creating a new func "getIdentities" to do
the same as getIdentity but without the logic that chooses an identity
by checking whether the username matches. This logic can't work for
emails.