getIdentity: it isn't always 2 identities that get returned
With the Limit field is set to 2, POST /identity/browse may return 4
identities. That's because /identity/browse applies the limit to each
identity provider. And since an AD or LDAP provider can access the local
identities, that's 2 identities coming from the local identities and 2
coming from the AD identity provider (for example).