contact-emails: address review comments
During the meeting with Luis, Russel, and Marcos, we talked about the
wording "programmer mistake". I doesn't sound very professional, so we
agreed on the following:

"this was not supposed to happen, please report to the developer team"

We also talked about the error message returned when no identity was
matched. The worry is that changing this error message may break users
who may be grepping on the error message.

This error message will rarely be returned. It appears when the user
forgot some letters at the end of the username. For example, they typed
jsmit instead of jsmith.

I still beleive that this is a corner case, and that changing this
message won't be breaking anyone.

That said, I agree with the fact that the error message isn't
descriptive enough.

I reverted back to the old message.