Update sample.capi.yaml
Where sanDNS was currently located it was being ignored.  it needed to be bumped out of the subject for the CSR.  Additionally, if it is located here it will do nothing and throw no errors.  I also and an extra \ into LocalMachine\\My.  in the current example it only had one which won't work and will confuse new users.